Refactor the test scripts
